FEAT #12471 TIME 0:10 fix tags migration + adding constraints

+++ migration/20.03/2003.sql
@@ -249,13 +249,18 @@ DO $$ BEGIN
       ALTER TABLE tags ADD COLUMN usage TEXT;
   END IF;
 END$$;
+
+DO $$ BEGIN
+  IF (SELECT count(attname) FROM pg_attribute WHERE attrelid = (SELECT oid FROM pg_class WHERE relname = 'tag_res')) = 1 THEN
+      DROP TABLE IF EXISTS resources_tags;
+      ALTER TABLE tag_res ADD COLUMN id serial NOT NULL;
+      ALTER TABLE tag_res RENAME TO resources_tags;
+  END IF;
+END$$;
 SELECT setval('tags_id_seq', (SELECT MAX(id) from tags));
 
 DROP TABLE IF EXISTS tags_entities;
 
-DROP TABLE IF EXISTS resources_tags;
-ALTER TABLE tag_res ADD COLUMN id serial NOT NULL;
-ALTER TABLE tag_res RENAME TO resources_tags;
 
 /* DOCTYPES */
 DO $$ BEGIN
